move SocksProxyNettyOptions back
