+++ atlasdb-container-test-utils/src/main/java/com/palantir/atlasdb/containers/SocksProxyNettyOptions.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright 2021 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package com.palantir.atlasdb.cassandra.proxy;
+package com.palantir.atlasdb.containers;
 
 import com.datastax.driver.core.NettyOptions;
 import io.netty.channel.socket.SocketChannel;
 import io.netty.handler.proxy.Socks5ProxyHandler;
 import java.net.SocketAddress;
 
-public final class SocksProxyNettyOptions extends NettyOptions {
+final class SocksProxyNettyOptions extends NettyOptions {
     private final SocketAddress proxyAddress;
 
-    public SocksProxyNettyOptions(SocketAddress proxyAddress) {
+    SocksProxyNettyOptions(SocketAddress proxyAddress) {
         this.proxyAddress = proxyAddress;
     }
